refactor(test): 테스트용 Inventory 분리
필드에 Anonymous class로 존재하던 Mock Inventory를 클래스로 분리합니다.
